refactor: remove unused Stripe controller files and update routes for webhook handling
