+++ internal/payment/handle_webhook.go
@@ -1,19 +1,21 @@
-package controllers
+package payment
 
 import (
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
+	"log"
 	"net/http"
 	"os"
+	"time"
 
+	"github.com/haikali3/gymbara-backend/internal/database"
 	"github.com/haikali3/gymbara-backend/pkg/utils"
 	"github.com/stripe/stripe-go/v81"
 	"github.com/stripe/stripe-go/v81/webhook"
 	"go.uber.org/zap"
 )
 
-func handleWebhook(w http.ResponseWriter, req *http.Request) {
+func HandleWebhook(w http.ResponseWriter, req *http.Request) {
 	const MaxBodyBytes = int64(65536)
 	bodyReader := http.MaxBytesReader(w, req.Body, MaxBodyBytes)
 	payload, err := ioutil.ReadAll(bodyReader)
@@ -27,7 +29,7 @@ func handleWebhook(w http.ResponseWriter, req *http.Request) {
 	// If you are using an endpoint defined with the API or dashboard, look in your webhook settings
 	// at https://dashboard.stripe.com/webhooks
 
-	endpointSecret := os.Getenv("STRIPE_SECRET_KEY")
+	endpointSecret := os.Getenv("STRIPE_WEBHOOK_SECRET")
 	signatureHeader := req.Header.Get("Stripe-Signature")
 	event, err := webhook.ConstructEvent(payload, signatureHeader, endpointSecret)
 	if err != nil {
@@ -36,43 +38,42 @@ func handleWebhook(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 
-	// Unmarshal the event data into an appropriate struct depending on its Type
 	switch event.Type {
-	case "customer.subscription.created":
+	case "customer.subscription.created", "customer.subscription.updated":
 		var subscription stripe.Subscription
 		err := json.Unmarshal(event.Data.Raw, &subscription)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error parsing webhook JSON: %v\n", err)
-			w.WriteHeader(http.StatusBadRequest)
+			http.Error(w, "Error parsing webhook JSON", http.StatusBadRequest)
 			return
+		}
+
+		// Calculate expiration date (1 month later)
+		expirationDate := time.Now().AddDate(0, 1, 0).Format("2006-01-02")
+
+		// Update database with new subscription details
+		_, err = database.DB.Exec(`
+			UPDATE Subscriptions 
+			SET stripe_subscription_id = $1, paid_date = NOW(), expiration_date = $2 
+			WHERE user_id = (SELECT id FROM Users WHERE email = $3)
+		`, subscription.ID, expirationDate, subscription.Customer.Email)
+
+		if err != nil {
+			log.Printf("DB error: %v", err)
+		}
+
+		// Also set is_premium = TRUE in Users table
+		_, err = database.DB.Exec(`
+			UPDATE Users SET is_premium = TRUE WHERE email = $1
+		`, subscription.Customer.Email)
+
+		if err != nil {
+			utils.Logger.Error("DB error.", zap.Error(err))
 		}
 		utils.Logger.Info("Subscription created.", zap.String("subscription_id", subscription.ID))
 		// Then define and call a func to handle the successful attachment of a PaymentMethod.
 		// handleSubscriptionCreated(subscription)
 
-	case "customer.subscription.deleted":
-		var subscription stripe.Subscription
-		err := json.Unmarshal(event.Data.Raw, &subscription)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error parsing webhook JSON: %v\n", err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		utils.Logger.Info("Subscription deleted.", zap.String("subscription_id", subscription.ID))
-		// Then define and call a func to handle the deleted subscription.
-		// handleSubscriptionCanceled(subscription)
-
-	case "customer.subscription.updated":
-		var subscription stripe.Subscription
-		err := json.Unmarshal(event.Data.Raw, &subscription)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error parsing webhook JSON: %v\n", err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		utils.Logger.Info("Subscription updated.", zap.String("subscription_id", subscription.ID))
-		// Then define and call a func to handle the successful attachment of a PaymentMethod.
-		// handleSubscriptionUpdated(subscription)
+	// Unmarshal the event data into an appropriate struct depending on its Type
 
 	case "customer.subscription.trial_will_end":
 		var subscription stripe.Subscription
@@ -97,6 +98,36 @@ func handleWebhook(w http.ResponseWriter, req *http.Request) {
 		utils.Logger.Info("Active entitlement summary updated.", zap.String("subscription_id", subscription.ID))
 		// Then define and call a func to handle active entitlement summary updated.
 		// handleEntitlementUpdated(subscription)
+
+	case "customer.subscription.deleted":
+		var subscription stripe.Subscription
+		err := json.Unmarshal(event.Data.Raw, &subscription)
+		if err != nil {
+			http.Error(w, "Error parsing webhook JSON", http.StatusBadRequest)
+			return
+		}
+
+		// Cancel subscription and set user as non-premium
+		_, err = database.DB.Exec(`
+			UPDATE Users SET is_premium = FALSE WHERE id = (SELECT user_id FROM Subscriptions WHERE stripe_subscription_id = $1)
+		`, subscription.ID)
+		if err != nil {
+			log.Printf("DB error: %v", err)
+		}
+
+		_, err = database.DB.Exec(`
+			UPDATE Subscriptions SET expiration_date = NOW() WHERE stripe_subscription_id = $1
+		`, subscription.ID)
+		if err != nil {
+			log.Printf("DB error: %v", err)
+		}
+
+		if err != nil {
+			log.Printf("DB error: %v", err)
+		}
+		utils.Logger.Info("Subscription deleted.", zap.String("subscription_id", subscription.ID))
+		// Then define and call a func to handle the deleted subscription.
+		// handleSubscriptionCanceled(subscription)
 
 	default:
 		utils.Logger.Warn("Unhandled event type.", zap.String("event_type", string(event.Type)))
